requirements.txt
boto3>=1.26.0
google-cloud-storage>=2.10.0
dropbox>=11.36.0
PyYAML>=6.0
tqdm>=4.65.0
python-dateutil>=2.8.2
schedule>=1.2.0
click>=8.1.3
loguru>=0.7.0
pytest>=7.0.0
pytest-cov>=4.0.0 